Log in screen and popups
Controls/how to play menu and the messages section pop up as a new window when their icons are clicked. Also main menu now takes you to log in screen which currently bypasses the narrative intro. This will be fixed soon as the intro text will appear in the form of a new message.